Remove abstract HeaderDecoder and HeaderEncoder
closes #17
We don't need to worry about custom implementations now that the type of
headers has been defined as Seq[(String,String)] in the http20 package.
