+++ http/src/main/scala/org/http4s/blaze/http/http20/HeaderEncoder.scala
@@ -1,15 +1,35 @@
 package org.http4s.blaze.http.http20
 
+import java.io.ByteArrayOutputStream
 import java.nio.ByteBuffer
+import java.nio.charset.StandardCharsets.US_ASCII
 
-/** This needs to contain the state of the header Encoder */
-trait HeaderEncoder {
+import com.twitter.hpack.Encoder
+import org.http4s.blaze.http.http20.Settings.DefaultSettings
+
+
+/** Simple Headers type for use in blaze and testing */
+final class HeaderEncoder(private var maxTableSize: Int = DefaultSettings.HEADER_TABLE_SIZE)
+{
+  require(maxTableSize <= DefaultSettings.HEADER_TABLE_SIZE, "Invalid initial table size")
+
+  private val encoder = new Encoder(getMaxTableSize)
+  private val os = new ByteArrayOutputStream(1024)
+
 
   /** Note that the default value is 4096 bytes */
-  def getMaxTableSize(): Int
+  def getMaxTableSize(): Int = maxTableSize
 
-  def setMaxTableSize(max: Int): Unit
+  /** If this is changed, the peer must be notified */
+  def setMaxTableSize(max: Int): Unit = {
+    maxTableSize = max
+    encoder.setMaxHeaderTableSize(os, max)
+  }
 
-  def encodeHeaders(hs: Seq[(String,String)]): ByteBuffer
+  def encodeHeaders(hs: Seq[(String, String)]): ByteBuffer = {
+    hs.foreach { case (k,v) => encoder.encodeHeader(os, k.getBytes(US_ASCII), v.getBytes(US_ASCII), false) }
+    val buff = ByteBuffer.wrap(os.toByteArray())
+    os.reset()
+    buff
+  }
 }
-
